refactor: separate different implementation functions, considering one missing pattern could have multiple implementations;
