+++ pygrinder/mar/__init__.py
@@ -1,0 +1,8 @@
+# Created by Wenjie Du <wenjay.du@gmail.com>
+# License: GLP-v3
+
+from .mar_logistic import mar_logistic
+
+__all__ = [
+    "mar_logistic",
+]
