+++ pygrinder/mar/mar_logistic.py
